Add lambda layer support.
Add the ability to import packages from lambda layers.

Lambda layers allow package dependencies to be installed via .zip file separate from the code deployment for the service. Currently, serverless-wsgi expects to either be handling packaging itself, or packaging to be handled via serverless-python-requirements built-in zip file deployment. If deploying dependencies via layers, serverless-wsgi might not find its own dependencies (Werkzeug not being found is the most obvious sign that something is wrong with packaging/deployment) or it might start but later imports of packages like boto3 or requests might load very old versions instead of the ones packaged with the service.

This adds two new settings in the custom: wsgi fields in serverless.yml

use_layers: boolean. Use layers for package deployment. Defaults to adding /opt/python to sys.path, which is the default for serverless-python-requirements if using 'layer: true'.

layer_path: override the default above with a different path, if the service is using nonstandard lambda layers.